Delete migrated MLMD Apis from metadata project
The relevant classes are available in frontend/mlmd.
