+++ frontend/package.json
@@ -5,7 +5,7 @@
   "dependencies": {
     "@material-ui/core": "^3.7.1",
     "@material-ui/icons": "^3.0.1",
-    "frontend": "kwasi/frontend#0289a2d",
+    "frontend": "kwasi/frontend#f732514",
     "grpc-web": "^1.0.7",
     "lodash.debounce": "^4.0.8",
     "lodash.groupby": "^4.6.0",
