+++ frontend/src/components/ModelInfo.test.tsx
@@ -1,8 +1,8 @@
+import {ArtifactProperties} from 'frontend';
 import * as React from 'react';
 import {ModelInfo} from './ModelInfo';
 import {shallow, ShallowWrapper} from 'enzyme';
 import {buildTestModel, stringValue, testModel} from '../TestUtils';
-import {ArtifactProperties} from '../lib/Api';
 
 describe('ModelInfo', () => {
   let tree: ShallowWrapper;
