+++ frontend/src/lib/Common.ts
@@ -1,0 +1,8 @@
+/** Format for a List request */
+export interface ListRequest {
+  filter?: string;
+  orderAscending?: boolean;
+  pageSize?: number;
+  pageToken?: string;
+  sortBy?: string;
+}
