+++ frontend/src/pages/ArtifactList.test.tsx
@@ -1,4 +1,5 @@
 import {
+  Api,
   Artifact,
   ArtifactType,
   GetArtifactsResponse,
@@ -8,7 +9,6 @@ import * as React from 'react';
 import ArtifactList from './ArtifactList';
 import {PageProps} from './Page';
 import {shallow, ShallowWrapper, ReactWrapper} from 'enzyme';
-import {Api} from '../lib/Api';
 import * as TestUtils from '../TestUtils';
 import {RoutePage} from '../components/Router';
 import CustomTable, {ExpandState} from '../components/CustomTable';
