+++ frontend/src/pages/ArtifactList.tsx
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import {ArtifactType, Artifact, GetArtifactsRequest, getArtifactTypes, getTypeName} from 'frontend';
+import {Api, ArtifactType, ArtifactProperties, ArtifactCustomProperties, Artifact, GetArtifactsRequest, getArtifactTypes, getTypeName} from 'frontend';
 import * as React from 'react';
 import CustomTable, {Column, Row, ExpandState, CustomRendererProps} from '../components/CustomTable';
 import {Page} from './Page';
@@ -22,9 +22,9 @@ import {ToolbarProps} from '../components/Toolbar';
 import {classes} from 'typestyle';
 import {commonCss, padding} from '../Css';
 import {rowCompareFn, rowFilterFn, groupRows, getExpandedRow, getResourceProperty} from '../lib/Utils';
-import {Api, ArtifactProperties, ArtifactCustomProperties, ListRequest} from '../lib/Api';
-import { RoutePage, RouteParams } from '../components/Router';
-import { Link } from 'react-router-dom';
+import {ListRequest} from '../lib/Common';
+import {RoutePage, RouteParams} from '../components/Router';
+import {Link} from 'react-router-dom';
 
 interface ArtifactListState {
   artifacts: Artifact[];
