+++ frontend/src/pages/ExecutionList.tsx
@@ -22,7 +22,7 @@ import { ToolbarProps } from '../components/Toolbar';
 import { classes } from 'typestyle';
 import { commonCss, padding } from '../Css';
 import { getResourceProperty, rowCompareFn, rowFilterFn, groupRows, getExpandedRow } from '../lib/Utils';
-import { Api, ListRequest, ExecutionProperties, ExecutionCustomProperties } from '../lib/Api';
+import { Api, ListRequest, ExecutionProperties, ExecutionCustomProperties } from '../lib/Common';
 import { Link } from 'react-router-dom';
 import { RoutePage, RouteParams } from '../components/Router';
